# frontend/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "MaintenanceCompleteRequest": {
        "properties": {
          "completed_on": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed On"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "title": "MaintenanceCompleteRequest",
        "type": "object"
      },
...
      "MaintenanceHistoryBoardResponse": {
        "properties": {
          "history_limit": {
            "title": "History Limit",
            "type": "integer"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/MaintenanceHistoryRowResponse"
            },
            "title": "Rows",
            "type": "array"
          }
        },
        "required": ["rows", "history_limit"],
        "title": "MaintenanceHistoryBoardResponse",
        "type": "object"
      },
...
      "MaintenanceLinkAppointmentRequest": {
        "properties": {
          "appointment_id": {
            "title": "Appointment Id",
            "type": "string"
          }
        },
        "required": ["appointment_id"],
        "title": "MaintenanceLinkAppointmentRequest",
        "type": "object"
      },
      "MaintenanceListFilter": {
        "enum": [
          "overdue",
          "due_soon",
          "needs_scheduling",
          "scheduled_upcoming",
          "active",
          "archived"
        ],
        "title": "MaintenanceListFilter",
        "type": "string"
      },
...
      "MaintenanceStatus": {
        "enum": ["active", "archived"],
        "title": "MaintenanceStatus",
        "type": "string"
      },
...
    }
  },
  "info": {
    "description": "Self-hosted personal planning platform API",
    "title": "Planforge",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/maintenance": {
      "get": {
        "operationId": "list_maintenance_endpoint_api_maintenance_get",
        "parameters": [
          {
            "in": "query",
            "name": "filter",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/MaintenanceListFilter"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Filter"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/MaintenanceStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MaintenanceResponse"
                  },
                  "title": "Response List Maintenance Endpoint Api Maintenance Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Maintenance Endpoint",
        "tags": ["maintenance"]
      },
      "post": {
        "operationId": "create_maintenance_endpoint_api_maintenance_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MaintenanceCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaintenanceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Maintenance Endpoint",
        "tags": ["maintenance"]
      }
    },
    "/api/maintenance/history-board": {
      "get": {
        "operationId": "maintenance_history_board_endpoint_api_maintenance_history_board_get",
        "parameters": [
          {
            "in": "query",
            "name": "history_limit",
            "required": false,
            "schema": {
              "default": 10,
              "maximum": 500,
              "minimum": 1,
              "title": "History Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaintenanceHistoryBoardResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Maintenance History Board Endpoint",
        "tags": ["maintenance"]
      }
    },
    "/api/maintenance/{maintenance_id}": {
      "get": {
        "operationId": "get_maintenance_endpoint_api_maintenance__maintenance_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "maintenance_id",
            "required": true,
            "schema": {
              "title": "Maintenance Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "history_limit",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 500,
              "minimum": 1,
              "title": "History Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaintenanceDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Maintenance Endpoint",
        "tags": ["maintenance"]
      },
      "patch": {
        "operationId": "update_maintenance_endpoint_api_maintenance__maintenance_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "maintenance_id",
            "required": true,
            "schema": {
              "title": "Maintenance Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MaintenanceUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaintenanceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Maintenance Endpoint",
        "tags": ["maintenance"]
      }
    },
    "/api/maintenance/{maintenance_id}/archive": {
      "post": {
        "operationId": "archive_maintenance_endpoint_api_maintenance__maintenance_id__archive_post",
        "parameters": [
          {
            "in": "path",
            "name": "maintenance_id",
            "required": true,
            "schema": {
              "title": "Maintenance Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaintenanceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Archive Maintenance Endpoint",
        "tags": ["maintenance"]
      }
    },
    "/api/maintenance/{maintenance_id}/clear-next-action": {
      "post": {
        "operationId": "clear_next_action_endpoint_api_maintenance__maintenance_id__clear_next_action_post",
        "parameters": [
          {
            "in": "path",
            "name": "maintenance_id",
            "required": true,
            "schema": {
              "title": "Maintenance Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaintenanceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Clear Next Action Endpoint",
        "tags": ["maintenance"]
      }
    },
    "/api/maintenance/{maintenance_id}/complete": {
      "post": {
        "operationId": "complete_maintenance_endpoint_api_maintenance__maintenance_id__complete_post",
        "parameters": [
          {
            "in": "path",
            "name": "maintenance_id",
            "required": true,
            "schema": {
              "title": "Maintenance Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/MaintenanceCompleteRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaintenanceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Complete Maintenance Endpoint",
        "tags": ["maintenance"]
      }
    },
    "/api/maintenance/{maintenance_id}/completions": {
      "post": {
        "operationId": "add_historical_completion_endpoint_api_maintenance__maintenance_id__completions_post",
        "parameters": [
          {
            "in": "path",
            "name": "maintenance_id",
            "required": true,
            "schema": {
              "title": "Maintenance Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MaintenanceHistoricalCompletionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaintenanceCompletionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Historical Completion Endpoint",
        "tags": ["maintenance"]
      }
    },
    "/api/maintenance/{maintenance_id}/completions/{completion_id}/correct": {
      "post": {
        "operationId": "correct_completion_endpoint_api_maintenance__maintenance_id__completions__completion_id__correct_post",
        "parameters": [
          {
            "in": "path",
            "name": "maintenance_id",
            "required": true,
            "schema": {
              "title": "Maintenance Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "completion_id",
            "required": true,
            "schema": {
              "title": "Completion Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MaintenanceCorrectCompletionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaintenanceCompletionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Correct Completion Endpoint",
        "tags": ["maintenance"]
      }
    },
    "/api/maintenance/{maintenance_id}/link-appointment": {
      "post": {
        "operationId": "link_appointment_endpoint_api_maintenance__maintenance_id__link_appointment_post",
        "parameters": [
          {
            "in": "path",
            "name": "maintenance_id",
            "required": true,
            "schema": {
              "title": "Maintenance Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MaintenanceLinkAppointmentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaintenanceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Link Appointment Endpoint",
        "tags": ["maintenance"]
      }
    },
    "/api/maintenance/{maintenance_id}/reschedule-appointment": {
      "post": {
        "operationId": "reschedule_linked_appointment_endpoint_api_maintenance__maintenance_id__reschedule_appointment_post",
        "parameters": [
          {
            "in": "path",
            "name": "maintenance_id",
            "required": true,
            "schema": {
              "title": "Maintenance Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MaintenanceScheduleAppointmentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaintenanceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reschedule Linked Appointment Endpoint",
        "tags": ["maintenance"]
      }
    },
    "/api/maintenance/{maintenance_id}/restore": {
      "post": {
        "operationId": "restore_maintenance_endpoint_api_maintenance__maintenance_id__restore_post",
        "parameters": [
          {
            "in": "path",
            "name": "maintenance_id",
            "required": true,
            "schema": {
              "title": "Maintenance Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaintenanceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Restore Maintenance Endpoint",
        "tags": ["maintenance"]
      }
    },
    "/api/maintenance/{maintenance_id}/schedule-appointment": {
      "post": {
        "operationId": "schedule_appointment_endpoint_api_maintenance__maintenance_id__schedule_appointment_post",
        "parameters": [
          {
            "in": "path",
            "name": "maintenance_id",
            "required": true,
            "schema": {
              "title": "Maintenance Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MaintenanceScheduleAppointmentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaintenanceDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Schedule Appointment Endpoint",
        "tags": ["maintenance"]
      }
    },
...
  }
}